% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grbasics.R
\name{gab2sq}
\alias{gab2sq}
\title{Convert an auxiliary block design to the treatment
information matrix of the corresponding  square array design}
\usage{
gab2sq(m, nc, abd)
}
\arguments{
\item{m}{Number of blocks m}

\item{nc}{Number of rows}

\item{abd}{Auxiliary block design}
}
\value{
cmat C matrix of corresponding square array design
}
\description{
Convert an auxiliary block design to the treatment
information matrix of the corresponding  square array design
}
