% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtbasics.R
\name{gacd}
\alias{gacd}
\title{Factors of an integer for calculating
the number of cyclic sets}
\usage{
gacd(x)
}
\arguments{
\item{x}{Integer}
}
\value{
factors Factors of the integer x
}
\description{
Factors of an integer for calculating
the number of cyclic sets
}
