% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtbasics.R
\name{gclass}
\alias{gclass}
\title{Equivalence classes for t-(k-1) <= 10 ONLY}
\usage{
gclass(nc, set)
}
\arguments{
\item{nc}{Number of rows/controls}

\item{set}{Set of designs with the same spacing}
}
\value{
ac  List of equivalence classes (not permuted)
}
\description{
Equivalence classes for t-(k-1) <= 10 ONLY
}
