% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grbasics.R
\name{grcmat}
\alias{grcmat}
\title{Calculates treatment information matrix C=X^t(I-Z(Z^TZ)^-)X of
a square array design}
\usage{
grcmat(m, xmat, zmat)
}
\arguments{
\item{m}{Dimension of square array m^2}

\item{xmat}{X matrix for treatments}

\item{zmat}{Z matrix for rows and columns}
}
\value{
cmat  C matrix for the square array
}
\description{
Calculates treatment information matrix C=X^t(I-Z(Z^TZ)^-)X of
a square array design
}
