% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grbasics.R
\name{grmets}
\alias{grmets}
\title{Calculates A metrics for a square array design}
\usage{
grmets(m, nc, ntest, ntreat, cmat)
}
\arguments{
\item{m}{Dimension of square array m^2}

\item{nc}{Number of controls}

\item{cmat}{C matrix of square array}
}
\value{
gvct  Variances for pairwise controls - test lines
}
\description{
Calculates A metrics for a square array design
}
