% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grbasics.R
\name{grnewzmat}
\alias{grnewzmat}
\title{Arrange Zmat0 according to the coordinates of the first row}
\usage{
grnewzmat(m, nc, gc, zmat0)
}
\arguments{
\item{m}{Dimension of square array m^2}

\item{nc}{Number of controls}

\item{gc}{Coordinates of controls in first row of the array}

\item{zmat0}{Basic Z matrix}
}
\value{
zmat Z matrix for coordinates gc
}
\description{
Arrange Zmat0 according to the coordinates of the first row
}
