% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grbasics.R
\name{grxmat}
\alias{grxmat}
\title{Calculates X matrix for square array}
\usage{
grxmat(m, nc)
}
\arguments{
\item{m}{Dimension of square array m^2}

\item{nc}{Number of controls}
}
\value{
xmat X matrix for square array
}
\description{
Calculates X matrix for square array
}
