% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grbasics.R
\name{grzmat0}
\alias{grzmat0}
\title{Calculates Z matrix for square array}
\usage{
grzmat0(m)
}
\arguments{
\item{m}{Dimension of square array m^2}
}
\value{
zmat   Basic Z matrix 2m x2m
}
\description{
Calculates Z matrix for square array
}
