% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtbasics.R
\name{tknum}
\alias{tknum}
\title{Finds number of cyclic sets in a cyclic design}
\usage{
tknum(nt, k)
}
\arguments{
\item{nt}{Number of parameters}

\item{k}{Number of row/controls}
}
\value{
Nsum Number of classes
}
\description{
Finds number of cyclic sets in a cyclic design
}
